rdcrdcrdc

+++ src/main/java/com/dit/hua/houseM/repositories/OwnerRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface OwnerRepository extends JpaRepository<Owner, Integer> {
+public interface OwnerRepository extends JpaRepository<Owner, Long> {
 
     Optional<Owner> findByUsername(String username);
