+++ src/main/java/com/dit/hua/houseM/repositories/RenterRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RenterRepository extends JpaRepository<Renter, Integer> {
+public interface RenterRepository extends JpaRepository<Renter, Long> {
     // Find renters by username
     List<Renter> findByUsername(String username);
